requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-pymongo==2.3.0
flask-bcrypt==0.7.1
flask-jwt-extended==4.2.3
face-recognition==1.3.0
python-dotenv==0.19.1